Simplify adding fakes to collection in collection.stub
