+++ lib/sinon/collection.js
@@ -11,12 +11,11 @@
 var sinonSpy = require("./spy");
 var sinonStub = require("./stub");
 var sinonMock = require("./mock");
-var valueToString = require("./util/core/value-to-string");
 var throwOnFalsyObject = require("./throw-on-falsy-object");
 var collectOwnMethods = require("./collect-own-methods");
+var stubNonFunctionProperty = require("./stub-non-function-property");
 
 var push = [].push;
-var hasOwnProperty = Object.prototype.hasOwnProperty;
 
 function getFakes(fakeCollection) {
     if (!fakeCollection.fakes) {
@@ -91,39 +90,22 @@ var collection = {
         return this.add(sinonSpy.apply(sinonSpy, arguments));
     },
 
-    stub: function stub(object, property, value) {
+    stub: function stub(object, property/*, value*/) {
         throwOnFalsyObject.apply(null, arguments);
 
         var isStubbingEntireObject = typeof property === "undefined" && typeof object === "object";
-
-        if (property) {
-            var original = object[property];
-
-            if (typeof original !== "function") {
-                if (!hasOwnProperty.call(object, property)) {
-                    throw new TypeError("Cannot stub non-existent own property " + valueToString(property));
-                }
-
-                object[property] = value;
-
-                return this.add({
-                    restore: function () {
-                        object[property] = original;
-                    }
-                });
-            }
-        }
+        var isStubbingNonFunctionProperty = property && typeof object[property] !== "function";
+        var stubbed = isStubbingNonFunctionProperty ?
+                        stubNonFunctionProperty.apply(null, arguments) :
+                        sinonStub.apply(null, arguments);
 
         if (isStubbingEntireObject) {
-            var col = this;
-            var stubbedObj = sinonStub.apply(null, arguments);
-
-            collectOwnMethods(stubbedObj).forEach(col.add.bind(this));
-
-            return stubbedObj;
+            collectOwnMethods(stubbed).forEach(this.add.bind(this));
+        } else {
+            this.add(stubbed);
         }
 
-        return this.add(sinonStub.apply(null, arguments));
+        return stubbed;
     },
 
     mock: function mock() {
